+++ main.py
@@ -1,6 +1,7 @@
 import telebot
 import feedback as fb
 import handlers
+from database import order
 
 
 bot = telebot.TeleBot('7367715020:AAEZortk_qDiDFA28I7LfAYnnbLsX1loE48')
@@ -39,6 +40,11 @@ def choose_dish(call):
     print('call', call.data)
     fb.fb_dish_selected(call, bot)
 
-
+@bot.callback_query_handler(func=lambda call: call.data.startswith('add_to_cart'))
+def handle_callback(call):
+    print(call.data)
+    dish_id = call.data.split(':')[1]
+    print('name', dish_id)
+    handlers.add_to_order(call.message, dish_id)
 
 bot.polling(none_stop=True)
